test(stubs): unify SketchUp stubs for async scan
