+++ tests/stubs/sketchup.rb
@@ -17,44 +17,68 @@ module Sketchup
     @extensions.clear
   end
 
-  class Model
-    attr_reader :selection, :observers
-
-    def initialize
-      @selection = Selection.new
-      @observers = []
-    end
-
-    def add_observer(obs)
-      @observers << obs
-    end
-
-    def remove_observer(obs)
-      @observers.delete(obs)
-    end
-  end
-
-  class Selection
-    attr_reader :observers
-
-    def initialize
-      @observers = []
-    end
-
-    def add_observer(obs)
-      @observers << obs
-    end
-
-    def remove_observer(obs)
-      @observers.delete(obs)
-    end
-  end
-
+  class ComponentInstance; end
+  class Group < ComponentInstance; end
   class ModelObserver; end
   class SelectionObserver; end
 
+  class Layer
+    def visible? = true
+    def name = ''
+  end
+
+  class Entities
+    def initialize(list)
+      @list = list
+    end
+
+    def to_a
+      @list
+    end
+  end
+
+  class Model
+    attr_reader :entities, :selection, :layers, :observers
+
+    def initialize(entities = [])
+      @entities  = Entities.new(entities)
+      @selection = Selection.new
+      @layers    = []
+      @observers = []
+    end
+
+    def add_observer(obs)
+      @observers << obs
+    end
+
+    def remove_observer(obs)
+      @observers.delete(obs)
+    end
+  end
+
+  class Selection < Entities
+    attr_reader :observers
+
+    def initialize
+      super([])
+      @observers = []
+    end
+
+    def add_observer(obs)
+      @observers << obs
+    end
+
+    def remove_observer(obs)
+      @observers.delete(obs)
+    end
+  end
+
   def self.active_model
     @active_model ||= Model.new
+  end
+
+  def self.active_model=(model)
+    @active_model = model
   end
 
   def self.temp_dir
@@ -64,5 +88,9 @@ module Sketchup
   def self.reset
     @active_model = Model.new
   end
+end
+
+module Geom
+  Z_AXIS = [0, 0, 1].freeze
 end
